fix(): refine general-card doc
Refs BRICK_STORE-2818
