+++ bricks/basic-bricks/src/interfaces/general-card.ts
@@ -1,0 +1,11 @@
+import { ButtonProps } from "antd/lib/button";
+
+export interface OperationButton {
+  // to listen for
+  id: string;
+  eventName: string;
+  configProps: ButtonProps & { icon?: string };
+  tooltip?: string;
+  text?: string;
+  needData?: boolean;
+}
