+++ bricks/basic-bricks/src/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from "./general-card";
